Renamed lock_device as tasmota_device, and separated Codes Generator from lock_device
